added SearchDropdown component , modified the navbar.jsx component
